working update

+++ config-2.sh
@@ -2,24 +2,26 @@
 echo "If you are running that means you have created a traning data set folder with all images and respective txt file of same name"
 # read -p "enter the folder name :- " dir
 dir="traning"
-# ls $dir
 cd ../darknet
-rm custom_data -R
-
+v=$(ls | grep custom_data)
+len=${#v}
+if [ $len == '11' ] ; then
+	echo "---Updating Custom Data Directory---"
+	rm custom_data -R
+else
+	echo "---Making Custom Data Directory---"
+fi
 mkdir custom_data
 cd custom_data
 mkdir images
+mkdir labels
 cd ../../Image-Processing
 
-# put p as percentage of test set
-# echo "Enter the Percentage of Test Data Set vs Traning Data Set "
-# p=50
 read -p "Enter the Percentage of Test Data Set( vs Traning Data Set ):- " p
+# the rest will be traning data set
 
-# the rest will be traning data ser
 cd $dir
 total=$(ls ima*.txt | sed 's/i/custom_data\\images\\i/' | wc -l)
-# echo $total
 r=$(($total*$p/100))
 
 echo "Test Data "$p'% = '$r 
@@ -28,12 +30,12 @@ r=$(($r+1))
 a=${r}',$d'
 b=${r}',$p'
 # ls ima*.txt | sed 's/i/custom_data\\image\\i/'
-ls ima*.txt | sed 's/i/custom_data\\images\\i/' | sed $a | sed 's/.txt/.jpg/g' >>../../darknet/custom_data/test.txt
+ls ima*.txt | sed 's/i/custom_data\/images\/i/' | sed $a | sed 's/.txt/.jpg/g' >>../../darknet/custom_data/test.txt
 echo "Test data created"
 # ls ima*.txt | sed 's/i/custom_data\\image\\i/' | sed '7,$d'
-# echo "hello"
+
 # ls ima*.txt | sed 's/i/custom_data\\image\\i/' | sed -n '7,$p'
-ls ima*.txt | sed 's/i/custom_data\\images\\i/' | sed -n $b | sed 's/.txt/.jpg/g' >>../../darknet/custom_data/train.txt
+ls ima*.txt | sed 's/i/custom_data\/images\/i/' | sed -n $b | sed 's/.txt/.jpg/g' >>../../darknet/custom_data/train.txt
 echo "Train data created"
 
 var1=$(ls im*.txt |  sed 's/.txt/.jpg/')
@@ -41,11 +43,11 @@ cp $var1 ../../darknet/custom_data/images
 echo "images copyed"
 
 var2=$(ls im*.txt)
-cp $var2 ../../darknet/custom_data/images
-echo "images data copyed"
+cp $var2 ../../darknet/custom_data/labels
+echo "images txt(data) file copyed"
 
 cat classes.txt >> ../../darknet/custom_data/custom.names
-echo "lables created"
+echo "labels created"
 
 var3=$(cat classes.txt | wc -l)
 echo "classes="$var3 >>../../darknet/custom_data/detector.data
@@ -66,7 +68,9 @@ mkdir ../../darknet/custom_data/cfg
 # 2.Yolo V2
 # 3.Yolo V3"
 # read -p "Choice :- " var4
-iter=4000
+iter=500
+# iter=4000
+# iter=502000
 var4=4
 if [ $var4 == '1' ]
 then
